# [Numpy] FFI: diag/diagonal/diag_indices_from