--- a/src/api/operator/numpy/np_matrix_op.cc
+++ b/src/api/operator/numpy/np_matrix_op.cc
@@ -98,7 +98,7 @@ MXNET_REGISTER_API("_npi.diag")
   attrs.op = op;
   SetAttrDict<op::NumpyDiagParam>(&attrs);
   NDArray* inputs[] = {args[0].operator mxnet::NDArray*()};
-  int num_inputs =1;
+  int num_inputs = 1;
   int num_outputs = 0;
   auto ndoutputs = Invoke(op, &attrs, num_inputs, inputs, &num_outputs, nullptr);
   *ret = ndoutputs[0];