… src/operator/numpy/np_matrix_op.cc